Add `position-at` & related `backfill-state` fns.
Add the `position-at` function to the the core namespace. Once users
call `setup!` they can use this function to find the position of the
random walk at time `t`.

Add the `backfill-state` function, which fills in the path of the given
state with newly-created bezier curve segments until the given time is
covered by a segment. It is used by `position-at` when the caller asks
for a time that is not yet covered by the state.