Add render-frame, API handler constructor.
The `render-frame` function is what renders the sketch into an offscreen
buffer, which we then pull a PNG from in order to send back to the HTTP
client. It needs to have the sketch applet itself, the nominal
dimensions, a configure function that sets any custom state on the
quil graphics object for each render (such as smoothing or a different
color mode), the draw function of the sketch, and finally the time and
scale at which to render the frame.

The `handler` constructor takes the state atom of the random walk and
all the sketch-related arguments of `render-frame`, and returns a ring
handler that can serve the state, reset the state via HTTP, and render
the sketch at the time specified by the client into a PNG image
response.