Add server management to core API.
Add an arity to the setup! function that takes all the information
needed in order to be able to render a frame of the sketch in response
to a web request. When all the necessary information has been supplied,
start a webserver that can serve the state, reset the state, and render
a frame out in response to web request.

To support this, add an atom to hold the current server instance, and
a pair of functions to start and stop it.

Add an option to the setup! function to set the port that the server is
started on. It defaults to 3000.