Add state resetting functionality.
When any preview image is clicked, that world's state is fetched, and
then all worlds are reset to have that state.

This is implemented in two parts: when an image is clicked, that world's
state is fetched asynchronously. When the state has been received,
a callback puts the state onto a core.async channel.

When the namespace is first loaded, a go-loop is defined that listens on
said channel for new states. When a new state is received, the current
list of worlds is retrieved from the Om state atom, and then the state
value is asynchronously posted to every world.