Load animation & curve defrecords in API ns.
In the API namespace, load the qutils.animation and qutils.curve
namespaces in order to evaluate the defrecord forms contained therein,
since if the record classes are not available, then deserializing the
submitted state will not work.

Also, use clojure 1.7.0-beta1, since in clojure 1.6.0, deserializing the
state never works once the API is running in many threads via Jetty (I
think this is because the wrong classloader is used, but I'm not sure).